+++ CHANGELOG.md
@@ -8,6 +8,8 @@ either pip or setup.py.
 Dependency Changes
 ------------------
 
+- No longer support pre-astropy versions of pyfits (now bundled in astropy
+  as astropy.io.fits). (#755)
 - Added LSSTDESC.Coord, which contains the functionality that used to be in
   GalSim as the Angle and CelestialCoord classes.  We moved it to a separate
   repo so people could more easily use this functionality without requiring all